fixing issie on orme env
(cherry picked from commit 746acec08287f254a7b432276d50fe1b0b1c8803)

+++ knowagecockpitengine/src/main/java/it/eng/spagobi/engine/chart/util/DataSetTransformer.java
@@ -846,7 +846,8 @@ public class DataSetTransformer {
 		if (drillOrder != null) {
 			for (String key : drillOrder.keySet()) {
 				Map<String, String> keyMapper = (Map<String, String>) drillOrder.get(key);
-				if (!keyMapper.get("orderColumn").equals("") && !keyMapper.get("orderColumn").equals(categorieColumns.get("column"))) {
+				if (!keyMapper.get("orderColumn").equals("") && !keyMapper.get("orderColumn").equals(categorieColumns.get("column"))
+						&& !drillOrder.containsKey(keyMapper.get("orderColumn"))) {
 					dataColumnsMapper.remove(keyMapper.get("orderColumn").toLowerCase());
 				}
 			}
